* develop:
  ATLAS-73 Rename unit-test target
  ATLAS-73 Fortran interface
  ATLAS-73 Fixes and unit-test
  ATLAS-73 Expose trans function to compute derivatives from spectral (untested)
  ATLAS-77 Spectral norms, Atlas interface
  Flush Log buffers when atlas_finalize() is called
  Fix sign comparison warnings
  Remove pragma's
  Fix sign comparison warnings

Conflicts:
	src/atlas/array/Array.h
	src/atlas/util/io/Gmsh.cc

+++ src/atlas/mesh/actions/BuildHalo.cc
@@ -534,7 +534,7 @@ public:
       buf.elem_part   [p][jelem] = elem_part(ielem);
       buf.elem_type   [p][jelem] = mesh.cells().type_idx(ielem);
       std::vector<double> crds(elem_nodes->cols(ielem)*2);
-      for( int jnode=0; jnode<elem_nodes->cols(ielem); ++jnode)
+      for(size_t jnode=0; jnode<elem_nodes->cols(ielem); ++jnode)
       {
         double crd[] = { lonlat( (*elem_nodes)(ielem,jnode),internals::LON) , lonlat( (*elem_nodes)(ielem,jnode),internals::LAT) };
         transform(crd,-1);
