+++ src/atlas/parallel/GatherScatter.h
@@ -319,7 +319,7 @@ void GatherScatter::scatter( parallel::Field<DATA_TYPE const> gfields[],
     throw eckit::SeriousBug("GatherScatter was not setup",Here());
   }
 
-  for( int jfield=0; jfield<nb_fields; ++jfield )
+  for(size_t jfield=0; jfield < nb_fields; ++jfield)
   {
     const int lvar_size = std::accumulate(lfields[jfield].var_shape.data(),lfields[jfield].var_shape.data()+lfields[jfield].var_rank,1,std::multiplies<int>());
     const int gvar_size = std::accumulate(gfields[jfield].var_shape.data(),gfields[jfield].var_shape.data()+gfields[jfield].var_rank,1,std::multiplies<int>());
@@ -330,7 +330,7 @@ void GatherScatter::scatter( parallel::Field<DATA_TYPE const> gfields[],
     std::vector<int> glb_displs(nproc);
     std::vector<int> glb_counts(nproc);
 
-    for (int jproc=0; jproc<nproc; ++jproc)
+    for (size_t jproc=0; jproc < nproc; ++jproc)
     {
       glb_counts[jproc] = glbcounts_[jproc]*gvar_size;
       glb_displs[jproc] = glbdispls_[jproc]*gvar_size;
@@ -370,7 +370,7 @@ void GatherScatter::pack_send_buffer( const parallel::Field<DATA_TYPE const>& fi
                                       const std::vector<int>& sendmap,
                                       DATA_TYPE send_buffer[] ) const
 {
-  const int sendcnt = sendmap.size();
+  const size_t sendcnt = sendmap.size();
 
   size_t ibuf = 0;
   const size_t send_stride = field.var_strides[0]*field.var_shape[0];
@@ -378,7 +378,7 @@ void GatherScatter::pack_send_buffer( const parallel::Field<DATA_TYPE const>& fi
   switch( field.var_rank )
   {
   case 1:
-    for( size_t p=0; p<sendcnt; ++p)
+    for(size_t p=0; p < sendcnt; ++p)
     {
       const size_t pp = send_stride*sendmap[p];
       for( size_t i=0; i<field.var_shape[0]; ++i )
@@ -389,7 +389,7 @@ void GatherScatter::pack_send_buffer( const parallel::Field<DATA_TYPE const>& fi
     }
     break;
   case 2:
-    for( size_t p=0; p<sendcnt; ++p)
+    for(size_t p=0; p < sendcnt; ++p)
     {
       const size_t pp = send_stride*sendmap[p];
       for( size_t i=0; i<field.var_shape[0]; ++i )
@@ -403,7 +403,7 @@ void GatherScatter::pack_send_buffer( const parallel::Field<DATA_TYPE const>& fi
     }
     break;
   case 3:
-    for( size_t p=0; p<sendcnt; ++p)
+    for(size_t p=0; p < sendcnt; ++p)
     {
       const size_t pp = send_stride*sendmap[p];
       for( size_t i=0; i<field.var_shape[0]; ++i )
@@ -520,7 +520,7 @@ void GatherScatter::gather( const array::ArrayView<DATA_TYPE,LRANK>& ldata,
                             array::ArrayView<DATA_TYPE,GRANK>& gdata,
                             const size_t root ) const
 {
-  if( ldata.shape(0) == parsize_ && gdata.shape(0) == glb_cnt(root) )
+  if( ldata.shape(0) == parsize_ && gdata.shape(0) == size_t(glb_cnt(root)) )
   {
     std::vector< parallel::Field<DATA_TYPE const> > lfields(1, parallel::Field<DATA_TYPE const>(ldata) );
     std::vector< parallel::Field<DATA_TYPE> >       gfields(1, parallel::Field<DATA_TYPE>(gdata) );
@@ -541,7 +541,7 @@ void GatherScatter::scatter( const array::ArrayView<DATA_TYPE,GRANK>& gdata,
                              array::ArrayView<DATA_TYPE,LRANK>& ldata,
                              const size_t root ) const
 {
-  if( ldata.shape(0) == parsize_ && gdata.shape(0) == glb_cnt(root) )
+  if( ldata.shape(0) == parsize_ && gdata.shape(0) == size_t(glb_cnt(root)) )
   {
     std::vector< parallel::Field<DATA_TYPE const> > gfields(1, parallel::Field<DATA_TYPE const>(gdata) );
     std::vector< parallel::Field<DATA_TYPE> >       lfields(1, parallel::Field<DATA_TYPE>(ldata) );
